refactor(bridge-cffi): add target-neutral Wasm support
