+++ baml_language/crates/bridge_cffi/Cargo.toml
@@ -21,21 +21,24 @@ rustls = [ "sys_native/rustls" ]
 bundle-http = [ "sys_native/bundle-http" ]
 
 [dependencies]
-baml_version = { workspace = true }
-bex_events = { workspace = true }
 bex_project = { workspace = true }
 bridge_ctypes = { workspace = true }
-sys_native = { workspace = true }
-sys_types = { workspace = true }
+sys_ops = { workspace = true }
 futures = { workspace = true }
 indexmap = { workspace = true }
+prost = { workspace = true }
+thiserror = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+baml_version = { workspace = true }
+bex_events = { workspace = true }
+sys_native = { workspace = true }
+sys_types = { workspace = true }
 libc = { workspace = true }
 once_cell = { workspace = true }
-prost = { workspace = true }
 serde_json = { workspace = true }
-thiserror = { workspace = true }
 tokio = { workspace = true, features = [ "rt", "rt-multi-thread" ] }
 vfs = { workspace = true }
 
 [package.metadata.ci]
-wasm_support = false
+wasm_support = true
